requirements.txt
# Master Orchestrator dependencies

spaCy>=1.4.0
python-dotenv>=0.19.0
GitPython>=3.1.30
aiohttp>=3.8.1
async-timeout>=4.0.2
tenacity>=8.0.1  # For retry logic

# For development and testing
pytest>=7.2.0
pytest-asyncio>=0.19.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.0.0